# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: aboutme API
  version: 0.1.0
  description: |
    `/api/v1` — the REST contract for the aboutme resume builder + hosted
    display service. This document is the single source of truth for the
    API. Contract changes update this file instead of redefining shapes
    elsewhere. Go, TypeScript, and Dart clients are generated from it.

    ## Response envelope

    Every response body is one of:

    - Success: `Envelope` — `{"data": ...}`.
    - Failure: `Error` — `{"error": {"code": "...", "message": "..."}}`.
      Both `code` and `message` are always present.

    ## Write safety (revisions, `If-Match`, idempotency)

    Every resume is guarded by a monotonically increasing `revision`
    (server-assigned bigint). Because 64-bit integers are not safely
    representable in JSON numbers for TS/Dart clients, `revision` is always
    serialized as a **string** — see the `Revision` schema.

    - **Optimistic concurrency.** Every mutation of an existing resume must
      carry the `If-Match` header in ETag form:
      `If-Match: "r<revision>"` (see the `IfMatch` parameter). Creating a
      resume has no prior revision and rejects `If-Match`. Authentication and
      session mutations do not use the resume precondition.
    - **Staleness is `412`, never `409`.** When the supplied `If-Match` does
      not match the resource's current revision, the server returns
      `412 Precondition Failed` with an `Error` body and the current
...
  license:
...
servers:
...
paths:
...
  /auth/google/start:
    get:
      operationId: getAuthGoogleStart
      tags: [auth]
      security: []
      summary: Begin "Sign in with Google"
      description: >-
        Begins a fresh OAuth2/OIDC login transaction (PKCE S256, an OIDC
        nonce) and redirects the browser to Google's own authorize
        endpoint. Sets the `__Host-oauth-tx` cookie (opaque transaction
        handle, 10-minute TTL) that the matching `GET
        /auth/google/callback` consumes.


        A `HEAD` request is rejected with `405`, not treated as a bodyless
        `GET`, because a successful request creates a database-backed
        transaction. `GET` accepts only the login purpose. A `link` or
        `reauth` purpose returns `405` with `Allow: POST` before a session
        lookup, database write, or cookie is set.
      parameters:
        - $ref: "#/components/parameters/AuthPurpose"
      responses:
        "302":
          description: Success — a redirect to Google's own authorize endpoint.
          headers:
            Location:
              schema:
                type: string
                format: uri
              example: https://accounts.google.com/o/oauth2/v2/auth?...
        "405":
          description: >-
            Either an unsupported method, including `HEAD`, or a `GET`
            carrying `?purpose=link` or `?purpose=reauth`. The latter
            purposes are served only by `POST` and return `Allow: POST`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/AuthStartRateLimited"
    post:
      operationId: postAuthGoogleStart
      tags: [auth]
      security:
        - sessionCookie: []
          csrfToken: []
      summary: Begin a Google link or reauthentication
      description: >-
        Begins an OAuth transaction for the CALLER's own already-
        authenticated account — `purpose=link` (attach this provider
        identity to that account) or `purpose=reauth` (refresh the
        session's `reauthenticated_at` against an ALREADY-linked
        identity) — and returns Google's authorize URL for the client to
        navigate to. Sets the same `__Host-oauth-tx` cookie (opaque
        transaction handle, 10-minute TTL) the matching `GET
        /auth/google/callback` consumes.


        The response deliberately does NOT redirect: a `302` answering a
        `fetch()` would be followed by the fetch, and the provider's
        consent screen has to be a real top-level navigation the client
        performs itself with the returned URL.


        Link and reauthentication starts are privileged account actions.
        They therefore use `POST` behind the normal session, CSRF token,
        and exact-Origin checks. `GET` refuses both purposes.
      parameters:
        - $ref: "#/components/parameters/AuthLinkPurpose"
      responses:
        "200":
          description: The authorize URL to navigate the browser to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStartResponse"
              example:
                data:
                  authorizeUrl: https://accounts.google.com/o/oauth2/v2/auth?...
        "400":
          description: >-
            `purpose` was absent or was something other than `link` or
            `reauth`. This route serves exactly those two — a login start
            is the `GET`'s job — and never substitutes a default, which
            would hand back an authorize URL for a flow the caller did not
            ask for.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: bad_request
                  message: purpose must be link or reauth
        "401":
          description: No valid `__Host-session` cookie. Clears the cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: session_required
                  message: a valid session is required
        "403":
          description: >-
            Either the CSRF chain rejected the request (missing/mismatched
            `Origin`, missing/wrong `csrfToken`) or — `purpose=link` only —
            the session's last full OAuth login is more than 15 minutes
            old. The reauth check runs BEFORE any transaction row is
            created, so a stale caller never gets a live transaction
            handle. `purpose=reauth` has no such gate: its whole point is
            to establish recency.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                csrf_rejected:
                  summary: Failed the CSRF chain
                  value:
                    error:
                      code: csrf_rejected
                      message: CSRF validation failed
                reauth_required:
                  summary: "purpose=link, stale reauth window"
                  value:
                    error:
                      code: reauth_required
                      message: recent reauthentication is required
        "429":
          $ref: "#/components/responses/AuthStartRateLimited"
  /auth/google/callback:
    get:
      operationId: getAuthGoogleCallback
      tags: [auth]
      security: []
      summary: Complete "Sign in with Google"
      description: >-
        Consumes the pending OAuth transaction (`__Host-oauth-tx`
        cookie), exchanges the authorization code (PKCE), verifies the ID
        token (issuer/audience/signature/expiry, plus the OIDC nonce),
        and resolves or creates the local user or attaches the identity to
        an already-authenticated one, depending on the transaction's own
        `purpose` (set at `/auth/google/start`):


        - `purpose=login` (default): requires `email_verified == true`,
          then resolves the identity — an already-linked identity signs
          straight back in with **no** email comparison at all; an unknown
          identity whose email already belongs to a DIFFERENT account
          rejects with `email_already_registered` and creates nothing (no
          automatic email merge, AC-AUTH-001); otherwise creates a new
          user + identity.

        - `purpose=link`/`reauth`: no email check at all. Re-derives the
          caller's own session from THIS request's `__Host-session`
          cookie (the anonymous `security: []` above describes only the
          `purpose=login` default) rather than from anything in the
          transaction itself. `link` attaches the identity to that caller
          (idempotent if already attached to them; `identity_already_linked`
          if claimed by someone else) and issues NO new session — the
          caller already has one. `reauth` refreshes that session's
          `reauthenticated_at` against an ALREADY-linked identity only.


        Always redirects. This is a top-level browser navigation, never a
        JSON response:


        - **Success**: `302` to the app's own origin (`/`) for
          `purpose=login`, or to `/app/settings/sessions` for
          `purpose=link`/`reauth` — with the `__Host-session`
          cookie set for `purpose=login` only.

        - **Rejected**: `302` to `/login?error=<code>` for `purpose=login`,
          or to `/app/settings/sessions?error=<code>` for `purpose=link`/
          `reauth`, with one of `OAuthCallbackErrorCode`'s codes and no
          session cookie set. The no-oracle contract means every internal
          rejection reason not explicitly listed
          collapses into the generic `auth_failed` code — the response
          never lets a caller distinguish which specific check failed.


        `__Host-oauth-tx` is cleared on every exit path, success or
        rejection alike.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: The provider's authorization code (RFC 6749 §4.1.2).
            Absent on a provider-signaled rejection (see `error` below).
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Must match the pending transaction's own `state` (RFC
            6749 §10.12) or the callback is rejected with `auth_failed`.
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Set by the PROVIDER itself (never this server) when the
            visitor declines consent — `error=access_denied` (RFC 6749
            §4.1.2.1), which this server maps to its own `cancelled` code.
      responses:
        "302":
          description: Always redirects — see the operation description for
            the success/rejected branches.
          headers:
            Location:
              schema:
                type: string
                format: uri
              example: https://aboutme.vn/login?error=auth_failed
...
  /auth/linkedin/callback:
    get:
      operationId: getAuthLinkedInCallback
      tags: [auth]
      security: []
      summary: Complete "Sign in with LinkedIn"
      description: >-
        Consumes the pending OAuth transaction (`__Host-oauth-tx`
        cookie), exchanges the authorization code (PKCE), verifies the ID
        token (issuer/audience/signature/expiry, plus the OIDC nonce), and
        resolves the local user per `purpose`, exactly like `GET
        /auth/google/callback` — see that
        operation's own description for the full `purpose=login` vs
        `purpose=link`/`reauth` behavior and redirect targets, with ONE
        LinkedIn-specific difference: `purpose=login`'s registration path
        (an UNKNOWN identity only — an existing identity's repeat login
        never re-checks email at all, matching every other provider)
        requires a verified email exactly like Google, EXCEPT LinkedIn's
        OIDC `email`/`email_verified` claims are OPTIONAL — an absent
        `email_verified` claim is never treated as true (AC-AUTH-002).
        `purpose=link`/`reauth` never checks email; linking an existing
        account remains allowed without a verified email.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: The provider's authorization code (RFC 6749 §4.1.2).
            Absent on a provider-signaled rejection (see `error` below).
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Must match the pending transaction's own `state` (RFC
            6749 §10.12) or the callback is rejected with `auth_failed`.
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Set by the PROVIDER itself when the visitor declines
            consent — `error=access_denied`, mapped to this server's own
            `cancelled` code.
      responses:
        "302":
          description: Always redirects — see the operation description for
            the success/rejected branches.
          headers:
            Location:
              schema:
                type: string
                format: uri
              example: https://aboutme.vn/login?error=auth_failed

  # ---- Session-authenticated JSON API (AC-AUTH-005) -----------------------
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
      description: >-
        Server-side opaque session token: `httpOnly`,
        `Secure`, `SameSite=Lax`, no `Domain` attribute (`__Host-`
        prefix). Never a JWT — a revocable, Postgres-backed session row.
        Every non-OAuth-redirect operation in the `auth` tag requires
        this.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: >-
        Synchronizer CSRF token:
        `base64url` (no padding) encoding of the authenticated session's
        own CSRF secret, obtained from `GET /me`'s `csrfToken` field —
        the only place it is ever returned — and required, alongside
        `sessionCookie`, on every mutating cookie-authenticated request
        (together with `SameSite=Lax` and an exact `Origin`/`Referer`
        check). Declared together with `sessionCookie` (AND, not OR) on
        every mutating operation in the `auth` tag.